added auth for user and pagination

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -19,7 +19,14 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        $projects = Project::all();
+        //$projects = Project::all();
+        $currentUserId = Auth::id();
+        //$projects = Project::where('user_id', $currentUserId)->paginate(3);
+        if ($currentUserId == 1) {
+            $projects = Project::paginate(3);
+        } else {
+            $projects = Project::where('user_id', $currentUserId)->paginate(3);
+        }
         return view('admin.projects.index', compact('projects'));
     }
 
@@ -61,7 +68,11 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        return view('admin.projects.show', compact('project'));
+        $currentUserId = Auth::id();
+        if ($currentUserId == $project->user_id || $currentUserId == 1) {
+            return view('admin.projects.show', compact('project'));
+        }
+        abort(403);
     }
 
     /**
@@ -69,6 +80,10 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         $types = Type::all();
         $technologies = Technology::all();
         return view('admin.projects.edit', compact('project', 'types', 'technologies'));
@@ -80,6 +95,10 @@ class ProjectController extends Controller
     public function update(UpdateProjectRequest $request, Project $project)
     {
 
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         $formData = $request->validated();
         $formData['slug'] = $project->slug;
         if ($project->title !== $formData['title']) {
@@ -110,6 +129,11 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         if ($project->image) {
             Storage::delete($project->image);
         }
