+++ app/Http/Controllers/Admin/TechnologyController.php
@@ -7,6 +7,7 @@ use App\Models\Technology;
 use App\Http\Requests\StoreTechnologyRequest;
 use App\Http\Requests\UpdateTechnologyRequest;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
 
 class TechnologyController extends Controller
 {
@@ -52,6 +53,10 @@ class TechnologyController extends Controller
      */
     public function edit(Technology $technology)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != 1) {
+            abort(403);
+        }
         return view('admin.technologies.edit', compact('technology'));
     }
 
@@ -76,6 +81,10 @@ class TechnologyController extends Controller
      */
     public function destroy(Technology $technology)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != 1) {
+            abort(403);
+        }
         $technology->delete();
         return to_route('admin.technologies.index')->with('message', "$technology->name eliminato con successo");
     }
